server: fix ci
